refactor: HttpParserクラスを作成
