+++ src/request/HttpParser.cpp
@@ -1,0 +1,19 @@
+#include "HttpParser.hpp"
+
+const std::string HttpParser::CRLF = "\r\n";
+const std::string HttpParser::WS = " \t";
+const std::string HttpParser::Delimiters = "\"(),/:;<=>?@[\\]{}";
+
+HttpParser::ParseErrorExeption::ParseErrorExeption(
+    const std::string& error_status, const std::string& reason)
+    : std::runtime_error(reason), error_status_(error_status) {}
+
+HttpParser::ParseErrorExeption::~ParseErrorExeption() throw() {}
+
+const std::string& HttpParser::ParseErrorExeption::getErrorStatus() const {
+  return error_status_;
+}
+
+HttpParser::HttpParser() {}
+
+HttpParser::~HttpParser() {}
